+++ RSS/DTOs/CreateGameDTO.cs
@@ -5,6 +5,6 @@ namespace RSS.DTOs
         public string Name { get; set; }
         public string GameType { get; set; }
         public string Status { get; set; }
-        public decimal PlayerCount { get; set; }
+        public int PlayerCount { get; set; }
     }
 }
